Fixed: No deflood for some command when length limit take action. 0.4.1.0RC6
git-svn-id: svn://svn.czdc.org/PtokaX@3 4c629181-62e4-40b4-9e0d-b9f719119ca2

+++ src/ClientTagManager.cpp
@@ -51,7 +51,7 @@ void ClientTagMan::Load() {
         TiXmlHandle cfg(&doc);
         TiXmlNode *clienttags = cfg.FirstChild("ClientTags").Node();
 		if(clienttags != NULL) {
-        	int i = 0;
+        	uint32_t i = 0;
 			TiXmlNode *child = NULL;
 			while((child = clienttags->IterateChildren(child)) != NULL) {
 				TiXmlNode *clienttag = child->FirstChild("ClientTag");
@@ -82,7 +82,7 @@ void ClientTagMan::Load() {
 		TiXmlElement clienttags("ClientTags");
 		const char* ClientPatts[] = { "++", "DCGUI", "DC", "oDC", "DC:PRO", "QuickDC", "LDC++", "R2++", "Goofy++", "PWDC++", "BDC++", "zK++" };
 		const char* ClientNames[] = { "DC++", "Valknut", "NMDC2", "oDC", "DC:PRO", "QuickDC", "LDC++", "R2++", "Goofy++", "PWDC++", "BDC++", "zK++" };
-		for(unsigned int i = 0; i < 12; i++) {
+		for(uint8_t i = 0; i < 12; i++) {
 			TiXmlElement clientpatt("ClientTag");
 			clientpatt.InsertEndChild(TiXmlText(ClientPatts[i]));
